Feat(auth): :pencil: auth implemented

+++ app/Http/Controllers/ListingController.php
@@ -35,7 +35,7 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
-        //$formFields['user_id'] = auth()->id();
+        $formFields['user_id'] = auth()->id();
 
         Listing::create($formFields);
 
@@ -50,9 +50,9 @@ class ListingController extends Controller
     // Update Listing Data
     public function update(ListingUpdateRequest $request, Listing $listing) {
         // Make sure logged in user is owner
-        // if($listing->user_id != auth()->id()) {
-        //     abort(403, 'Unauthorized Action');
-        // }
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
 
         $formFields = $request->all();
 
@@ -68,16 +68,16 @@ class ListingController extends Controller
     // Delete Listing
     public function destroy(Listing $listing) {
         // Make sure logged in user is owner
-        // if($listing->user_id != auth()->id()) {
-        //     abort(403, 'Unauthorized Action');
-        // }
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
 
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully');
     }
 
     // Manage Listings
-    // public function manage() {
-    //     return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
-    // }
+    public function manage() {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
+    }
 }
